split up store
